+++ decidim-initiatives/app/commands/decidim/initiatives/admin/update_initiative.rb
@@ -30,6 +30,7 @@ module Decidim
             current_user,
             attributes
           )
+          notify_initiative_is_extended if @notify_extended
           broadcast(:ok, initiative)
         rescue ActiveRecord::RecordInvalid
           broadcast(:invalid, initiative)
@@ -55,9 +56,21 @@ module Decidim
             attrs[:signature_start_time] = form.signature_start_time
             attrs[:signature_end_time] = form.signature_end_time
             attrs[:offline_votes] = form.offline_votes
+
+            @notify_extended = true if form.signature_end_time != initiative.signature_end_time &&
+                                       form.signature_end_time > initiative.signature_end_time
           end
 
           attrs
+        end
+
+        def notify_initiative_is_extended
+          Decidim::EventsManager.publish(
+            event: "decidim.events.initiatives.initiative_extended",
+            event_class: Decidim::Initiatives::ExtendInitiativeEvent,
+            resource: initiative,
+            recipient_ids: initiative.followers.pluck(:id)
+          )
         end
       end
     end